Use correctly sized Windows x64 data type for setting console text colour (to remove a warning)

+++ travel-expense-system/src/ui.cpp
@@ -91,7 +91,7 @@ void ui::setTextColour(Colour colour)
     // No-op on other operating systems.
     #if _WIN32
 
-    long win32ColourFlag = 0;
+    unsigned short win32ColourFlag = 0;
 
     switch (colour)
     {
